requirements.txt
# Employee Attrition Prediction System - Dependencies
# Python 3.9+ recommended

# Data & ML
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
xgboost>=1.7.0
shap>=0.42.0

# Experiment tracking & tuning
mlflow>=2.8.0
optuna>=3.2.0

# API & Dashboard
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.28.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Config & utils
pyyaml>=6.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Optional for production
joblib>=1.2.0

# Testing
pytest>=7.0.0
